return killshot to before I changed it and extend it for survival. Update parser with new extended class
